+++ Runtime/Shapes/Shapes.Mesh.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using Unity.Collections;
 using UnityEngine;
 #if !UNITY_2021_1_OR_NEWER
@@ -39,18 +38,9 @@ namespace Vertx.Debugging
 				if (Mesh == null)
 					return;
 
-#if !UNITY_2020_1_OR_NEWER
-				if (!Mesh.isReadable)
-				{
-					Debug.LogWarning($"{Mesh} must be marked as Read/Write to use {nameof(MeshNormals)}.");
-					return;
-				}
-#endif
-
 				Matrix4x4 matrix = Matrix;
 				var unscaledMatrix = Matrix4x4.TRS(matrix.MultiplyPoint(Vector3.zero), matrix.rotation, Vector3.one);
 
-#if UNITY_2020_1_OR_NEWER
 				var array = Mesh.AcquireReadOnlyMeshData(Mesh);
 
 				for (int i = 0; i < array.Length; i++)
@@ -63,16 +53,6 @@ namespace Vertx.Debugging
 					for (int j = 0; j < meshData.vertexCount; j++)
 						new Arrow(matrix.MultiplyPoint(vertices[j]), unscaledMatrix.MultiplyVector(normals[j]) * ArrowLength).Draw(ref commandBuilder, color, duration);
 				}
-#else
-				using (ListPool<Vector3>.Get(out List<Vector3> vertices))
-				using (ListPool<Vector3>.Get(out List<Vector3> normals))
-				{
-					Mesh.GetVertices(vertices);
-					Mesh.GetNormals(normals);
-					for (int i = 0; i < vertices.Count; i++)
-						new Arrow(matrix.MultiplyPoint(vertices[i]), unscaledMatrix.MultiplyVector(normals[i]) * ArrowLength).Draw(ref commandBuilder, color, duration);
-				}
-#endif
 			}
 #endif
 		}